feat(ci): repo quality audit — ruff expansion, mypy scope, shim check, full lane
Expands the quality gate from two lanes (fast/ci) to three (fast/ci/full)
and tightens linting, type coverage, and dead-code scanning across the
codebase.

Tooling additions (requirements-dev.txt, .python-version, .pre-commit-config.yaml):
- pytest-cov + coverage for coverage tracking
- pre-commit with ruff and mypy hooks for local feedback before push
- vulture for dead-code scanning (informational, not a hard gate)
- pip-audit for dependency vulnerability scanning (full lane)
- radon for complexity metrics on auto_cycle.py (informational)
- .python-version = 3.11 for pyenv/asdf/Copilot signal

pyproject.toml changes:
- ruff: expanded select from [E,F] → [E,F,I,B,UP,SIM,C4,RUF]; suppress
  SIM117 (nested with), RUF001/002/003 (intentional em dashes), E501
- mypy: expanded enforced scope from 3 modules to 5 (+ plugins, config)
- mypy: new overrides block suppressing [misc] on plugin adapter decorators
- mypy: new overrides block ignoring errors in shim redirect modules
- coverage: run/report config added

New CI scripts:
- scripts/ci/check_shim_integrity.py: two-phase check — (1) prevents logic
  drift back into shim files, (2) prevents canonical code from forming new
  import deps on shim adapter paths; wired into fast lane
- scripts/ci/run_vulture.py: vulture runner with plugin-system whitelist
  (adapters are dynamically loaded, never statically imported); exits 0

Quality gate (run_repo_quality_gate.py):
- fast lane: adds check_shim_integrity.py
- full lane (new): ci + coverage + vulture + radon on auto_cycle.py + pip_audit
- CI workflow: --lane full on push to main, --lane ci on PRs

Ruff / mypy fixes across src/ scripts/ tests/:
- B006 noqa for intentional closure-capture defaults in auto_cycle.py
- B904 (raise from None) in sweep_and_report.py
- RUF012 ClassVar annotations on mutable class attrs in test files
- SIM102 merge in reporting/warnings.py
- SIM115 noqa on lock file open() (must stay open for fcntl)
- cw_adapter: import from canonical lab.plugins.core path, not shim
- pgd_adapter: _validate_positive_int value param widened to int | float
- base_defense.postprocess: signature narrowed to list (matches all impls)
- yolo_adapter (plugins/core): predict param typed as Iterable[Path]
- plugins/catalog: build_plugin_inventory return type fixed to dict[str, Any]
- _BaseCDogAdapter: _stage_name field added to base to fix attr-defined errors
- dpc_unet_wrapper: return type narrowed to dict[str, float | bool]

Tests updated to match new fast-lane command count (4 commands, not 3).
486 tests pass, 0 failures.

+++ src/lab/plugins/core/attacks/square_adapter.py
@@ -82,8 +82,8 @@ class SquareAttack:
             queries_used = 0
             for step in range(self.n_queries):
                 p = self.p_init * (0.5 ** (step / max(self.n_queries, 1)))
-                sq_h = max(1, int(math.ceil(p * h)))
-                sq_w = max(1, int(math.ceil(p * w)))
+                sq_h = max(1, math.ceil(p * h))
+                sq_w = max(1, math.ceil(p * w))
 
                 top = int(
                     torch.randint(
